Allow consecutive whitespace characters
Scribe's default behaviour is to take something like `<p>hey &nbsp;
&nbsp; there</p>` and return `<p>hey there</p>` which is not what we
want. We want users to be able to add spaces between words and have them
render on the user site.

Remove `replaceNbspCharsFormatter` from the Scribe formatter chain.
There is no way to _configure_ this in our version, only to remove it.